requirements.txt
numpy
matplotlib
scipy
torch
pylint
pytest
torch
pyserial